Added logger setup call
