+++ pipeline/scraper/scrape.js
@@ -21,49 +21,11 @@ var fs = require('fs');
 var fs_promise = require('fs-readdir-promise');
 var readline = require('readline');
 
-
+//Unix pipeline setup
 var unix = require('unix-dgram');
 var client = unix.createSocket('unix_dgram');
 
-//Logging mechanisim for script
-const EMERG = 0,
-    ALERT = 1,
-    CRIT = 2,
-    ERR = 3,
-    WARNING = 4,
-    NOTICE = 5,
-    INFO = 6,
-    DEBUG = 7;
-
-var prefixes = ['<0>', '<1>', '<2>', '<3>', '<4>', '<5>', '<6>', '<7>'];
-
-var logger = {
-    //console.log(prefixes[INFO], txt);
-    info: function(txt) {
-        console.log(prefixes[INFO], txt);
-    },
-    err: function(txt) {
-        console.log(prefixes[ERR], txt);
-    },
-    alert: function(txt) {
-        console.log(prefixes[ALERT], txt);
-    },
-    crit: function(txt) {
-        console.log(prefixes[CRIT], txt);
-    },
-    warn: function(txt) {
-        console.log(prefixes[WARNING], txt);
-    },
-    notice: function(txt) {
-        console.log(prefixes[NOTICE], txt);
-    },
-    debug: function(txt) {
-        console.log(prefixes[DEBUG], txt);
-    }
-}
-logger.info("Logger initialised");
-
-
+var logger = require("./logger.js");
 
 gplay.search({term: "facebook"}).then(console.log,console.log);
 
@@ -305,10 +267,7 @@ wordStashFiles.then(files => {
                         return scrapeWord(word).then(function(appsData) {
 
                             logger.info("Search apps total: "+ appsData.length);
-
-                            logger.info("Search apps total: "+ appsData.length);
-
-                            var r = Promise.resolve();
+                              var r = Promise.resolve();
 
                             appsData.forEach(app => {
